+++ RELEASE_NOTES.md
@@ -33,7 +33,10 @@ Prolog adapter and integration files
 * UPDATED: Scryer Prolog adapter file with support for reporting warning and
 error line numbers.
 
-* UPDATED: Require Trealla Prolog 1.8.61 or later version.
+* UPDATED: Require Trealla Prolog 1.8.66 or later version.
+
+* UPDATED: Trealla Prolog adapter file with experimental code for expanding
+`use_module/1` Prolog directives.
 
 Library
 -------